% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Example_Data.r
\docType{data}
\name{example_data}
\alias{example_data}
\title{Example data}
\format{
A data frame with 1000 rows and 6 variables:
\describe{
  \item{normal_data_1}{Normal Distributed Data 1}
  \item{normal_data_2}{Normal Distributed Data 1}
  \item{nonnormal_data_1}{non-Normal Distributed Data 1}
  \item{nonnormal_data_2}{non-Normal Distributed Data 2}
  \item{group_two}{Categorical Data that has two groups}
  \item{group_three}{Categorical Data that has three groups}
}
}
\usage{
example_data
}
\description{
Example data
}
\keyword{datasets}
