% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonpar_test.r
\name{nonpar_test}
\alias{nonpar_test}
\title{Nonparametric Hypothesis Testing}
\usage{
nonpar_test(
  df,
  test = "sign",
  group_var = NA,
  num_var = NA,
  paired = F,
  exact = F,
  mu_values = rep(0, 1),
  alternative = "two.sided",
  threshold = 0.05,
  method = "none"
)
}
\arguments{
\item{df}{Dataframe}

\item{test}{string to indicate if user want to mean test or variance test (not available for one sample test). Value could be either "sign" or "variance", the default value is sign.}

\item{group_var}{a column name indicate the group variable}

\item{num_var}{a list of column names that contains non-normal data}

\item{paired}{a Boolean value indicate if its a paired test, default value is False}

\item{exact}{a Boolean value indicate if its a exact test, default value is False}

\item{alternative}{a variable indicate the hypothesis testing, default is "two.sided"}

\item{threshold}{the threshold for Test, default value is 0.05}

\item{method}{correction method for Dunn's test, default value is "none"}

\item{mu}{a list of variable that indicate the null hypothesis assumption value, default is a list of 0}
}
\value{
a list where the first element contains the column names of normal data, the second element contains the column name for those data are not normal
}
\description{
a function that perform non-parametric hypothesis test result for all non-normal numeric columns in a dataframe
}
\examples{
nonpar_test(dataframe)
}
\author{
Yushu Zou
}
