% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_normality.r
\name{test_normality}
\alias{test_normality}
\title{Testing Normality}
\usage{
test_normality(df, threshold = 0.05)
}
\arguments{
\item{df}{Dataframe}

\item{threshold}{the threshold for Shapiro-Wilk test, default value is 0.05}
}
\value{
a list where the first element contains the column names of normal data, the second element contains the column name for those data are not normal
}
\description{
a function that test the normality of all numeric column in one dataframe
}
\examples{
set.seed(123)
n <- 1000
df <- data.frame(
normal1 = rnorm(n, mean = 50, sd = 10),    # Normally distributed
normal2 = rnorm(n, mean = 100, sd = 20),   # Normally distributed
uniform = runif(n, min = 0, max = 50),     # Uniform distribution (non-normal)
poisson = rpois(n, lambda = 20),           # Poisson distribution (non-normal)
exponential = rexp(n, rate = 0.1)          # Exponential distribution (non-normal)
)
test_normality(df)
}
\author{
Yushu Zou
}
