% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorical_test.r
\name{categorical_test}
\alias{categorical_test}
\title{Categorical Hypothesis Testing}
\usage{
categorical_test(
  df,
  group_var = NA,
  cat_var = NA,
  alternative = "two.sided",
  threshold = 0.05,
  correct = T
)
}
\arguments{
\item{df}{Dataframe}

\item{group_var}{a column name indicate the group variable}

\item{cat_var}{a list of column names that contains categorical data}

\item{alternative}{a variable indicate the hypothesis testing, default is "two.sided"}

\item{threshold}{the threshold for Test, default value is 0.05}

\item{correct}{a Boolean variable to indicate whether the chi-square test need correction}
}
\value{
a list where the first element contains the column names of normal data, the second element contains the column name for those data are not normal
}
\description{
a function that perform hypothesis test result for all categorical columns in a dataframe
}
\examples{
set.seed(123)
df <- data.frame(
 group = factor(rep(c("Group1", "Group2"), each = 50)), # Two groups
 category1 = factor(sample(c("Cat1", "Cat2", "Cat3"), 100, replace = TRUE)),
 category2 = factor(sample(c("TypeA", "TypeB", "TypeC"), 100, replace = TRUE))
)
categorical_test(df, group_var = "group")
}
\author{
Yushu Zou
}
