% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_normality.r
\name{test_normality}
\alias{test_normality}
\title{Testing Normality}
\usage{
test_normality(df, threshold = 0.05)
}
\arguments{
\item{df}{Dataframe}

\item{threshold}{the threshold for Shapiro-Wilk test, default value is 0.05}
}
\value{
a list where the first element contains the column names of normal data, the second element contains the column name for those data are not normal
}
\description{
a function that test the normality of all numeric column in one dataframe
}
\examples{
test_normality(example_data)
}
\author{
Yushu Zou
}
